+++ packages/components/src/layout/TotalLayout/TotalLayoutContainer/TotalLayoutContainer.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { isFunction } from 'lodash';
 import { Box } from '../../Box';
 import { Stack } from '../../Stack';
 import { TotalLayoutContainerStyles } from './TotalLayoutContainer.styles';
@@ -32,15 +33,15 @@ const TotalLayoutContainer = ({ Header, Footer, scrollRef, children }) => {
   }, [scrollRef?.current, handleScroll]);
 
   return (
-    <Box id="TotalLayout" style={{ height: '100vh' }}>
+    <Box id="TotalLayout" className={classes.root}>
       <Stack fullWidth fullHeight direction="column">
         <Box className={classes.header} noFlex>
-          <Header style={{ position: 'fixed', top: 0, height: '72px' }} />
+          {isFunction(Header) ? <Header /> : Header}
         </Box>
-        <Box style={{ overflow: 'hidden' }}>{children}</Box>
+        {children}
         {!!Footer && (
           <Box className={classes.footerContainer} noFlex>
-            {Footer}
+            {isFunction(Footer) ? <Footer /> : Footer}
           </Box>
         )}
       </Stack>
@@ -52,9 +53,9 @@ TotalLayoutContainer.defaultProps = {
   Header: <div />,
 };
 TotalLayoutContainer.propTypes = {
-  Header: PropTypes.node,
-  Footer: PropTypes.node,
-  children: PropTypes.node,
+  Header: PropTypes.any,
+  Footer: PropTypes.any,
+  children: PropTypes.any,
   scrollRef: PropTypes.object,
 };