+++ packages/components/src/layout/TotalLayout/TotalLayoutStepContainer/TotalLayoutStepContainer.styles.js
@@ -1,13 +1,17 @@
 import { createStyles } from '@mantine/styles';
 
-const TotalLayoutStepContainerStyles = createStyles((theme) => ({
+const TotalLayoutStepContainerStyles = createStyles((theme, { hasFooter }) => ({
   root: {},
   stepContainer: {
     padding: '24px 0 0 0 ',
     width: '928px',
-    height: '100vh',
+    height: '100%',
   },
-  formContainer: { backgroundColor: 'white', padding: '24px' },
+  formContainer: {
+    backgroundColor: 'white',
+    padding: '24px',
+    paddingBottom: hasFooter ? '100px' : '24px',
+  },
   stepName: {
     fontSize: '23px',
     fontWeight: 500,