+++ packages/components/src/layout/TotalLayout/index.js
@@ -2,3 +2,5 @@ export * from './TotalLayout';
 export * from './TotalLayoutHeader/TotalLayoutHeader';
 export * from './TotalLayoutStepContainer/TotalLayoutStepContainer';
 export * from './TotalLayout.constants';
+export * from './TotalLayoutContainer/TotalLayoutContainer';
+export * from './TotalLayoutFooterContainer/TotalLayoutFooterContainer';