+++ packages/components/src/navigation/VerticalStepper/Step/Step.styles.js
@@ -8,7 +8,7 @@ const getStepHeight = (isButton, isText, isActivity) => {
   return 44;
 };
 
-export const StepStyles = createStyles(
+const StepStyles = createStyles(
   (
     theme,
 
@@ -23,25 +23,25 @@ export const StepStyles = createStyles(
       isVisited,
       isCompleted,
       isChildActive,
-    }
+    },
   ) => {
     const stepperTheme = theme.other.stepper ?? {};
     const stepHeight = getStepHeight(isButton, isText, isActivity);
     const childHeight = childRange ? 32 * (childRange[1] - childRange[0] + 1) : 0;
     const getLabelColor = () => {
       let color = stepperTheme.content?.color?.default;
-      if(isActivity) {
-        theme.colors.interactive01
-      } else if(isActive) {
+      if (isActivity) {
+        theme.colors.interactive01;
+      } else if (isActive) {
         color = stepperTheme.content?.color?.active;
-      } else if(isCompleted) {
+      } else if (isCompleted) {
         color = stepperTheme.content?.color?.completed;
       }
       return color ?? '#000';
-    }
+    };
     return {
       root: {
-        paddingLeft: !isChild ? 24 : 40,
+        paddingLeft: !isChild ? 0 : 40,
         paddingTop: !isChild ? (isText ? 6 : isActivity ? 4 : 12) : 4,
         paddingBottom: isText ? 6 : isActivity ? 4 : 12,
         paddingRight: 8,
@@ -55,7 +55,7 @@ export const StepStyles = createStyles(
         display: isChild && !showChild ? 'none' : 'flex',
         paddingRight: 8,
         position: 'relative',
-        paddingLeft: isChild ? 40 : 24,
+        paddingLeft: isChild ? 40 : 0,
         paddingTop: isChild ? 4 : 12,
         paddingBottom: 12,
       },
@@ -65,7 +65,7 @@ export const StepStyles = createStyles(
         paddingRight: 8,
         alignItems: 'center',
         position: 'relative',
-        paddingLeft: 24,
+        paddingLeft: 0,
         paddingTop: 12,
         paddingBottom: 12,
       },
@@ -132,5 +132,7 @@ export const StepStyles = createStyles(
         zIndex: -1,
       },
     };
-  }
+  },
 );
+
+export { StepStyles };